+++ src/js/_config/whatsNew.ts
@@ -1,15 +1,24 @@
 export const whatsNew = [
+  // {
+  //   version: '0.1.0',
+  //   title: 'Jellyfin support',
+  //   date: 'September 2025',
+  //   body: `<p>Jellyfin users can now enjoy their music libraries in Chromatix too!</p>
+  //   <p>Right now you can only log into one service at a time, but multiple account support and quick switching will arrive in another update soon.</p>`,
+  //   imageDark: 'promo-009-dark',
+  //   imageLight: 'promo-009-light',
+  // },
   {
-    version: '0.1.0',
-    title: 'Jellyfin support',
-    date: 'July 2025',
-    body: `<p>Jellyfin users can now enjoy their music libraries in Chromatix too!</p>
-    <p>Right now you can only log into one service at a time, but multiple account support and quick switching will arrive in another update soon.</p>`,
-    imageDark: 'promo-009-dark',
-    imageLight: 'promo-009-light',
+    version: '0.51.0',
+    title: 'Full screen player',
+    date: 'August 2025',
+    body: `<p>Chromatix has a brand new full screen player mode.</p>
+    <p>Click the new full screen icon in the bottom toolbar to try it out. You have access to all your music controls, and can also click the settings icon for customisation options.</p>`,
+    imageDark: 'promo-010-dark',
+    imageLight: 'promo-010-light',
   },
   {
-    version: '0.1.0',
+    version: '0.41.0',
     title: 'Artist track view',
     date: 'May 2025',
     body: `<p>A new track view has been added to artist pages, allowing you to view, play and shuffle all tracks by an artist in one go.</p>
@@ -18,7 +27,7 @@ export const whatsNew = [
     imageLight: 'promo-009-light',
   },
   {
-    version: '0.1.0',
+    version: '0.37.0',
     title: 'Column options',
     date: 'April 2025',
     body: `<p>List view columns can now be customised, allowing you to show and hide the fields you want. New fields have also been added, such as audio codec and bitrate.</p>
@@ -27,7 +36,7 @@ export const whatsNew = [
     imageLight: 'promo-008-light',
   },
   {
-    version: '0.1.0',
+    version: '0.32.0',
     title: 'Search & folders',
     date: 'Feb 2025',
     body: `<p>A search bar has been added to the main menu, letting you search for artists, albums, tracks, playlists, and more.</p>